Complete BigDecimal migration for all test utilities and data generators
- Fix BaseControllerTest, TestDataBuilder, and ValidationTestUtils for BigDecimal ficoScore
- Update BaseIntegrationTest and performance TestDataGenerator with BigDecimal support
- Fix all TestDataGenerator classes to return BigDecimal values instead of Integer
- Update CobolComparisonUtils method signature to accept BigDecimal parameters
- Fix TestDataGeneratorTest assertions with proper BigDecimal boundaries
- Comprehensive BigDecimal migration ensures COBOL COMP-3 precision parity across entire test suite

+++ backend/src/test/java/com/carddemo/integration/BaseIntegrationTest.java
@@ -195,7 +195,7 @@ public abstract class BaseIntegrationTest extends AbstractBaseTest {
                 .phoneNumber1("555-123-4567")
                 .ssn("999999999")
                 .dateOfBirth(LocalDate.of(1990, 1, 1))
-                .ficoScore(750)
+                .ficoScore(BigDecimal.valueOf(750))
                 .build();
     }
